Group check tests by object
