+++ tests/checks/test_baseline.py
@@ -7,7 +7,7 @@ IS_UNIX = not helpers.is_platform("windows")
 # General
 
 
-def test_validate():
+def test_validate_baseline():
     report = validate("data/table.csv")
     assert report.valid
 
@@ -29,14 +29,14 @@ def test_validate_invalid():
 # Stats
 
 
-def test_validate_stats_hash():
+def test_validate_baseline_stats_hash():
     hash = "6c2c61dd9b0e9c6876139a449ed87933"
     report = validate("data/table.csv", stats={"hash": hash})
     if IS_UNIX:
         assert report.task["valid"]
 
 
-def test_validate_stats_hash_invalid():
+def test_validate_baseline_stats_hash_invalid():
     hash = "6c2c61dd9b0e9c6876139a449ed87933"
     report = validate("data/table.csv", stats={"hash": "bad"})
     if IS_UNIX:
@@ -45,14 +45,14 @@ def test_validate_stats_hash_invalid():
         ]
 
 
-def test_validate_stats_hash_md5():
+def test_validate_baseline_stats_hash_md5():
     hash = "6c2c61dd9b0e9c6876139a449ed87933"
     report = validate("data/table.csv", stats={"hash": hash})
     if IS_UNIX:
         assert report.task["valid"]
 
 
-def test_validate_stats_hash_md5_invalid():
+def test_validate_baseline_stats_hash_md5_invalid():
     hash = "6c2c61dd9b0e9c6876139a449ed87933"
     report = validate("data/table.csv", stats={"hash": "bad"})
     if IS_UNIX:
@@ -61,14 +61,14 @@ def test_validate_stats_hash_md5_invalid():
         ]
 
 
-def test_validate_stats_hash_sha1():
+def test_validate_baseline_stats_hash_sha1():
     hash = "db6ea2f8ff72a9e13e1d70c28ed1c6b42af3bb0e"
     report = validate("data/table.csv", hashing="sha1", stats={"hash": hash})
     if IS_UNIX:
         assert report.task["valid"]
 
 
-def test_validate_stats_hash_sha1_invalid():
+def test_validate_baseline_stats_hash_sha1_invalid():
     hash = "db6ea2f8ff72a9e13e1d70c28ed1c6b42af3bb0e"
     report = validate("data/table.csv", hashing="sha1", stats={"hash": "bad"})
     if IS_UNIX:
@@ -77,14 +77,14 @@ def test_validate_stats_hash_sha1_invalid():
         ]
 
 
-def test_validate_stats_hash_sha256():
+def test_validate_baseline_stats_hash_sha256():
     hash = "a1fd6c5ff3494f697874deeb07f69f8667e903dd94a7bc062dd57550cea26da8"
     report = validate("data/table.csv", hashing="sha256", stats={"hash": hash})
     if IS_UNIX:
         assert report.task["valid"]
 
 
-def test_validate_stats_hash_sha256_invalid():
+def test_validate_baseline_stats_hash_sha256_invalid():
     hash = "a1fd6c5ff3494f697874deeb07f69f8667e903dd94a7bc062dd57550cea26da8"
     report = validate("data/table.csv", hashing="sha256", stats={"hash": "bad"})
     if IS_UNIX:
@@ -96,14 +96,14 @@ def test_validate_stats_hash_sha256_invalid():
         ]
 
 
-def test_validate_stats_hash_sha512():
+def test_validate_baseline_stats_hash_sha512():
     hash = "d52e3f5f5693894282f023b9985967007d7984292e9abd29dca64454500f27fa45b980132d7b496bc84d336af33aeba6caf7730ec1075d6418d74fb8260de4fd"
     report = validate("data/table.csv", hashing="sha512", stats={"hash": hash})
     if IS_UNIX:
         assert report.task["valid"]
 
 
-def test_validate_stats_hash_sha512_invalid():
+def test_validate_baseline_stats_hash_sha512_invalid():
     hash = "d52e3f5f5693894282f023b9985967007d7984292e9abd29dca64454500f27fa45b980132d7b496bc84d336af33aeba6caf7730ec1075d6418d74fb8260de4fd"
     report = validate("data/table.csv", hashing="sha512", stats={"hash": "bad"})
     if IS_UNIX:
@@ -115,13 +115,13 @@ def test_validate_stats_hash_sha512_invalid():
         ]
 
 
-def test_validate_stats_bytes():
+def test_validate_baseline_stats_bytes():
     report = validate("data/table.csv", stats={"bytes": 30})
     if IS_UNIX:
         assert report.task["valid"]
 
 
-def test_validate_stats_bytes_invalid():
+def test_validate_baseline_stats_bytes_invalid():
     report = validate("data/table.csv", stats={"bytes": 40})
     assert report.task.error.get("rowPosition") is None
     assert report.task.error.get("fieldPosition") is None
@@ -131,13 +131,13 @@ def test_validate_stats_bytes_invalid():
         ]
 
 
-def test_validate_stats_rows():
+def test_validate_baseline_stats_rows():
     report = validate("data/table.csv", stats={"rows": 2})
     if IS_UNIX:
         assert report.task["valid"]
 
 
-def test_validate_stats_rows_invalid():
+def test_validate_baseline_stats_rows_invalid():
     report = validate("data/table.csv", stats={"rows": 3})
     assert report.task.error.get("rowPosition") is None
     assert report.task.error.get("fieldPosition") is None
